[build] - Add support for building with ROCm 6.3 sources (#1189)
ROCr and ROCt are now in the same repo, we no longer use build_roct.sh.
Updated manifests to use 6.3 branches.
Updated ROCr patch to remove the specialized OpenMP cmake variables. We
no longer need this as ROCt is built with ROCr.
Updated build_supp.sh for 6.3 and removed openclicdloader.